Refactored the backend systems into their own folder.
