+++ Ayria/Source/Appmain.cpp
@@ -7,119 +7,6 @@
 #include <Stdinclude.hpp>
 #include <Global.hpp>
 
-namespace Backend
-{
-    static void setThreadname(std::string_view Name)
-    {
-        #pragma pack(push, 8)
-        using THREADNAME_INFO = struct { DWORD dwType; LPCSTR szName; DWORD dwThreadID; DWORD dwFlags; };
-        #pragma pack(pop)
-
-        __try
-        {
-            THREADNAME_INFO Info{ 0x1000, Name.data(), 0xFFFFFFFF };
-            RaiseException(0x406D1388, 0, sizeof(Info) / sizeof(ULONG_PTR), (ULONG_PTR *)&Info);
-        } __except (EXCEPTION_EXECUTE_HANDLER) {}
-    }
-
-    static DWORD __stdcall Graphicsthread(void *)
-    {
-        // UI-thread, boost our priority.
-        SetThreadPriority(GetCurrentThread(), THREAD_PRIORITY_HIGHEST);
-
-        // Name this thread for easier debugging.
-        if constexpr (Build::isDebug) setThreadname("Ayria_Graphics");
-
-        // Initialize the subsystems.
-        // TODO(tcn): Initialize pluginmenu.
-        Overlay_t Ingameconsole({}, {});
-        Console::Overlay::Createconsole(&Ingameconsole);
-
-        // Optional console for developers, runs its own thread.
-        if (std::strstr(GetCommandLineA(), "-DEVCON")) Console::Windows::Createconsole(GetModuleHandleW((NULL)));
-
-        // Main loop, runs until the application terminates or DLL unloads.
-        std::chrono::high_resolution_clock::time_point Lastframe{};
-        while (true)
-        {
-            // Track the frame-time, should be less than 33ms at worst.
-            const auto Thisframe{ std::chrono::high_resolution_clock::now() };
-            const auto Deltatime = std::chrono::duration<float>(Thisframe - Lastframe).count();
-            Lastframe = Thisframe;
-
-            // Notify all elements about the frame.
-            Ingameconsole.doFrame(Deltatime);
-            Console::Windows::doFrame();
-
-            // Log frame-average every 5 seconds.
-            if constexpr (Build::isDebug)
-            {
-                static std::array<uint32_t, 256> Timings{};
-                static float Elapsedtime{};
-                static size_t Index{};
-
-                Timings[Index % 256] = 1000000 * Deltatime;
-                Elapsedtime += Deltatime;
-                Index++;
-
-                if (Elapsedtime >= 5.0f)
-                {
-                    const auto Sum = std::reduce(std::execution::par_unseq, Timings.begin(), Timings.end());
-                    Debugprint(va("Average framerate: %5.2f FPS %5lu us", 1000000.0 / (Sum / 256), Sum / 256));
-                    Elapsedtime = 0;
-                }
-            }
-
-            // For reference, my workstations E5 hits 200FPS while laptops i5 gets 50FPS. Cap to 60.
-            std::this_thread::sleep_until(Lastframe + std::chrono::milliseconds(16));
-        }
-
-        return 0;
-    }
-    static DWORD __stdcall Backgroundthread(void *)
-    {
-        // Mainly IO bound thread, might as well sleep a lot.
-        SetThreadPriority(GetCurrentThread(), THREAD_PRIORITY_LOWEST);
-
-        // Name this thread for easier debugging.
-        if constexpr (Build::isDebug) setThreadname("Ayria_Background");
-
-        // Main loop, runs until the application terminates or DLL unloads.
-        while (true)
-        {
-            // Notify the subsystems about a new frame.
-            Auxiliary::Updatenetworking();
-            Clientinfo::doFrame();
-
-            // TODO(tcn): Get async-job requests and process them here.
-        }
-
-        return 0;
-    }
-
-    static void Initialize()
-    {
-        // Set SSE to behave properly, MSVC seems to be missing a flag.
-        _mm_setcsr(_mm_getcsr() | 0x8040); // _MM_FLUSH_ZERO_ON | _MM_DENORMALS_ZERO_ON
-
-        // Initialize subsystems that plugins may need.
-        Clientinfo::API_Initialize();
-        Auxiliary::API_Initialize();
-        Social::API_Initialize();
-        Clientinfo::Initialize();
-
-        // Default network groups.
-        Auxiliary::Joinmessagegroup(Auxiliary::Generalport);
-        Auxiliary::Joinmessagegroup(Auxiliary::Pluginsport);
-        Auxiliary::Joinmessagegroup(Auxiliary::Matchmakeport);
-        Auxiliary::Joinmessagegroup(Auxiliary::Fileshareport);
-
-        // Workers.
-        CreateThread(NULL, NULL, Graphicsthread, NULL, STACK_SIZE_PARAM_IS_A_RESERVATION, NULL);
-        CreateThread(NULL, NULL, Backgroundthread, NULL, STACK_SIZE_PARAM_IS_A_RESERVATION, NULL);
-    }
-}
-
 // Some applications do not handle exceptions well.
 static LONG __stdcall onUnhandledexception(PEXCEPTION_POINTERS Context)
 {
